Add the functions from PathUtils to Utils
