--- srcs/app/utils/PathUtils.cpp
@@ -1,24 +1,0 @@
-#include <unistd.h>
-#include <string>
-
-std::string	PathExtension(const std::string &path) {
-	const std::size_t extension_position = path.rfind(".");
-	if (extension_position == std::string::npos || extension_position < 2) {
-		return "";
-	}
-	const std::size_t	last_dir_position = path.rfind("/");
-	if (last_dir_position == std::string::npos) {
-		return path.substr(extension_position + 1);
-	}
-	if (last_dir_position < extension_position - 1) {
-		const std::string last_path_part = path.substr(last_dir_position + 1);
-		if (last_path_part != "." && last_path_part != "..") {
-			return path.substr(extension_position + 1);
-		}
-	}
-	return "";
-}
-
-bool	IsExecutable(const std::string &path) {
-	return access(path.c_str(), X_OK) == 0;
-}
